refactor: 역할 분리를 위한 컴포넌트 리팩토링
